requirements.txt
coloredlogs==14.0
python-json-logger==0.1.11
python-dateutil==2.8.1
pyfiglet==0.7
termcolor==1.1.0
colorama==0.4.3
tqdm==4.43.0
pdfminer.six==20200124
xlrd==1.2.0
XlsxWriter==1.2.8
python-telegram-bot==12.4.2
pandas==1.0.3
